Stub API behavior in tests
In order to avoid hitting the real API during tests execution, these
changes specify a hardcoded string in PlanetTest and use it as
a stubbed value.

This is achieved by adding a constructor to Planet.java that takes
an Http instance as a parameter. This is the instance that will
call for get() method inside of getClimate().

Then, we change the tests to inject the stub to Planet when
instantiating it using the constructor (Dependency Injection).