Verify params passed to URL in PlanetTest
Add a guard class to the stub to ensure that search term
is passed as param when calling getClimate().

Now, the stub is close to a mock (an object that mimics the behavior
of a testable part of the application in a controlled way and
verifies whether or not it received the expected params).

This has been done mannually in these changes. However, there is
a general-purpose mock tool that could be used for situations like
these. It is called Mockito.

Ref: https://site.mockito.org/