Use Mockito DI tools
Refactor tests to use Mockito Dependency Inject tools.

Create a mock instance of an Http object through @Mock annotation
and inject it into a target instance variable that represents
a Planet object (this one takes the @InjectMocks annotation). Initialize
these mocks inside a @Before method.

This refactor also enables the constructor in Planet class to be
removed. Actually, this constructor it was never needed for making
a Planet object work, it had been only added due to test purposes
during f9c4c71.

When using Mockito DI tools, mocking can be done at the field level,
eliminating the need for a constructor for this.